+++ packages/react-ui/storybook/stories/atoms/Text-field.stories.js
@@ -3,7 +3,12 @@ import { Grid, InputAdornment, TextField } from '@mui/material';
 import React from 'react';
 import PasswordField from '../../../src/components/atoms/PasswordField';
 import Typography from '../../../src/components/atoms/Typography';
-import { Container, Label } from '../../utils/storyStyles';
+import {
+  Container,
+  DocContainer,
+  DocTextHighlight,
+  Label
+} from '../../utils/storyStyles';
 
 const options = {
   title: 'Atoms/Text Field',
@@ -537,8 +542,22 @@ const PasswordFieldTemplate = ({
   return (
     <Grid container direction='column' spacing={6}>
       <Grid item>
+        <DocContainer severity='warning'>
+          Use{' '}
+          <DocTextHighlight variant='inherit' component='span'>
+            {'<PasswordField />'}
+          </DocTextHighlight>{' '}
+          instead of{' '}
+          <DocTextHighlight variant='inherit' component='span'>
+            {'<TextField type="password" />'}
+          </DocTextHighlight>{' '}
+          This component build the show / hide content logic on top of <i>TextField</i>{' '}
+          Mui component.
+        </DocContainer>
+      </Grid>
+      <Grid item>
         <Container>
-          <Label variant='body2'>{'Default'}</Label>
+          <Label variant='body2'>{'TextField'}</Label>
           <TextField
             {...rest}
             label='Password'
@@ -551,7 +570,7 @@ const PasswordFieldTemplate = ({
       </Grid>
       <Grid item>
         <Container>
-          <Label variant='body2'>{'Custom component'}</Label>
+          <Label variant='body2'>{'PasswordField'}</Label>
           <PasswordField
             {...rest}
             label='Password'