+++ packages/react-ui/storybook/stories/organisms/AppBar.stories.js
@@ -3,6 +3,7 @@ import { Avatar, Chip, Grid, IconButton, Menu, MenuItem } from '@mui/material';
 import { CheckCircleOutline } from '@mui/icons-material';
 import AppBar from '../../../src/components/organisms/AppBar/AppBar';
 import Typography from '../../../src/components/atoms/Typography';
+import { DocContainer, DocLink, DocTextHighlight } from '../../utils/storyStyles';
 
 const options = {
   title: 'Organisms/AppBar',
@@ -89,6 +90,32 @@ const CustomTemplate = (args) => {
   );
 };
 
+const DocTemplate = () => {
+  return (
+    <DocContainer severity='warning'>
+      We have our own{' '}
+      <DocLink href='https://github.com/CartoDB/carto-react/blob/master/packages/react-ui/src/components/organisms/AppBar/AppBar.js'>
+        AppBar
+      </DocLink>{' '}
+      to build the basic structure and styles on top of <i>Mui AppBar</i> and normalize
+      size and position of common elements like logo, texts, avatar and burger menu icon.
+      <Typography variant='inherit' color='inherit' mt={2}>
+        So, instead of Mui AppBar, you should use this one:{' '}
+        <DocTextHighlight variant='inherit' component='span'>
+          react-ui/src/components/organisms/AppBar
+        </DocTextHighlight>
+      </Typography>
+      <Typography variant='inherit' color='inherit' mt={2}>
+        For external use:{' '}
+        <DocTextHighlight variant='inherit' component='span'>
+          {'import { AppBar } from "@carto/react-ui";'}
+        </DocTextHighlight>
+        .
+      </Typography>
+    </DocContainer>
+  );
+};
+
 const commonArgs = {
   brandText: 'CARTO',
   secondaryText: 'Some text',
@@ -100,3 +127,5 @@ Basic.args = { ...commonArgs };
 
 export const Composition = CustomTemplate.bind({});
 Composition.args = { ...commonArgs };
+
+export const Guide = DocTemplate.bind({});