+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Add onClick event handler to GoogleMap component [#747](https://github.com/CartoDB/carto-react/pull/747)
+
 ## 2.2
 
 ### 2.2.1 (2023-07-21)